Update website and docs links
The URLs for website and docs have changed. Despite the efforts to have
redirects, some links don't work anymore at all ("Getting started",
"Working with clusters").
Use the new URLs and also place only the getting started guide as docs
links because it contains all other links. Instead of copying them we
can tell the contents to expect to be covered in the getting started
guide. The "Working with clusters" section seems to be gone and even
if it's meant to point to the docs section "Setup and Operation →
"Managing Clusters" I don't see why we should call it out here because
the getting started guide gives a better introduction with links to
relevant docs sections.